added events SDK

+++ src/events.ts
@@ -1,42 +1,57 @@
 import matter from 'gray-matter';
 import fs from 'node:fs/promises';
 import { join } from 'node:path';
-import { glob } from 'glob';
 import { dirname } from 'node:path';
-import { copy, remove } from 'fs-extra';
+import { copyDir, findFileById, getFiles, searchFilesForId, versionExists } from './internal/utils';
+import type { Event } from './types';
 
-type Badge = {
-  content: string;
-  backgroundColor: string;
-  textColor: string;
-};
+/**
+ * Returns an events from EventCatalog
+ * @param id - The id of the event to retrieve
+ * @param version - Optional id of the version to get
+ * @returns
+ */
+export const getEvent =
+  (directory: string) =>
+  async (id: string, version?: string): Promise<Event> => {
+    const file = await findFileById(directory, id, version);
 
-type Event = {
-  id: string;
-  name: string;
-  version: string;
-  summary: string;
-  owners?: string[];
-  badges?: Badge[];
-  // Path to the schema file.
-  schemaPath?: string;
-  // Markdown content of the event
-  markdown: string;
-};
+    if (!file)
+      throw new Error(
+        `No event found for the given id: ${id}` + (version ? ` and version ${version}` : '')
+      );
+
+    const { data, content } = matter.read(file);
+
+    return {
+      ...data,
+      markdown: content.trim(),
+    } as Event;
+  };
 
 /**
  * Adds an event to EventCatalog
  *
- * @param path - The path to your event, e.g. `/Inventory/InventoryAdjusted`
  * @param event - The event to write
+ * @param options - Optional options to write the event
  *
  */
-export const writeEvent = (directory: string) => async (path: string, event: Event) => {
-  const { markdown, ...frontmatter } = event;
-  const document = matter.stringify(markdown.trim(), frontmatter);
-  await fs.mkdir(join(directory, path), { recursive: true });
-  await fs.writeFile(join(directory, path, 'index.md'), document);
-};
+export const writeEvent =
+  (directory: string) =>
+  async (event: Event, options: { path: string } = { path: '' }) => {
+    // Get the path
+    const path = options.path || `/${event.id}`;
+    const exists = await versionExists(directory, event.id, event.version);
+
+    if (exists) {
+      throw new Error(`Failed to write event as the version ${event.version} already exists`);
+    }
+
+    const { markdown, ...frontmatter } = event;
+    const document = matter.stringify(markdown.trim(), frontmatter);
+    await fs.mkdir(join(directory, path), { recursive: true });
+    await fs.writeFile(join(directory, path, 'index.md'), document);
+  };
 
 /**
  * Remove an event to EventCatalog (modeled on the standard POSIX rm utility)
@@ -46,38 +61,6 @@ export const writeEvent = (directory: string) => async (path: string, event: Eve
  */
 export const rmEvent = (directory: string) => async (path: string) => {
   await fs.rm(join(directory, path), { recursive: true });
-};
-
-const getFiles = async (pattern: string) => {
-  try {
-    const files = await glob(pattern, { ignore: 'node_modules/**' });
-    return files;
-  } catch (error) {
-    throw new Error(`Error finding files: ${error}`);
-  }
-};
-
-const searchFilesForId = async (files: string[], id: string, version?: string) => {
-  const idRegex = new RegExp(`^id:\\s*['"]?${id}['"]?\\s*$`, 'm');
-  const versionRegex = new RegExp(`^version:\\s*['"]?${version}['"]?\\s*$`, 'm');
-
-  const matches = await Promise.all(
-    files.map(async (file) => {
-      const content = await fs.readFile(file, 'utf-8');
-      const hasIdMatch = content.match(idRegex);
-
-      // Check version if provided
-      if (version && !content.match(versionRegex)) {
-        return undefined;
-      }
-
-      if (hasIdMatch) {
-        return file;
-      }
-    })
-  );
-
-  return matches.filter(Boolean).filter((file) => file !== undefined);
 };
 
 /**
@@ -117,28 +100,13 @@ export const versionEvent = (directory: string) => async (id: string) => {
   const eventDirectory = dirname(file);
   const { data: { version = '0.0.1' } = {} } = matter.read(file);
   const targetDirectory = join(eventDirectory, 'versioned', version);
-  const tmpDirectory = join(directory, 'tmp');
 
   await fs.mkdir(targetDirectory, { recursive: true });
-  await fs.mkdir(tmpDirectory, { recursive: true });
 
-  // Copy everything over
-  await copy(eventDirectory, tmpDirectory, {
-    overwrite: true,
-    filter: (src) => {
-      return !src.includes('versioned');
-    },
+  // Copy the event to the versioned directory
+  await copyDir(directory, eventDirectory, targetDirectory, (src) => {
+    return !src.includes('versioned');
   });
-
-  await copy(tmpDirectory, targetDirectory, {
-    overwrite: true,
-    filter: (src) => {
-      return !src.includes('versioned');
-    },
-  });
-
-  // Remove the tmp directory
-  await fs.rm(tmpDirectory, { recursive: true });
 
   // Remove all the files in the root of the resource as they have now been versioned
   await fs.readdir(eventDirectory).then(async (resourceFiles) => {
@@ -151,3 +119,32 @@ export const versionEvent = (directory: string) => async (id: string) => {
     );
   });
 };
+
+/**
+ * Adds a file to the given event
+ * @param id - The id of the event to add the file to
+ * @param file - File contents to add including the content and the file name
+ * @param version - Optional version of the event to add the file to
+ * @returns
+ */
+export const addFileToEvent =
+  (directory: string) =>
+  async (id: string, file: { content: string; fileName: string }, version?: string) => {
+    const pathToEvent = await findFileById(directory, id, version);
+    if (!pathToEvent) throw new Error('Cannot find directory to write file to');
+    const contentDirectory = dirname(pathToEvent);
+    await fs.writeFile(join(contentDirectory, file.fileName), file.content);
+  };
+
+/**
+ * Adds a schema to the given event
+ * @param id - The id of the event to add the schema to
+ * @param schema - Schema contents to add including the content and the file name
+ * @param version - Optional version of the event to add the schema to
+ * @returns
+ */
+export const addSchemaToEvent =
+  (directory: string) =>
+  async (id: string, schema: { schema: string; fileName: string }, version?: string) => {
+    await addFileToEvent(directory)(id, { content: schema.schema, fileName: schema.fileName }, version);
+  };
